Update the main.js and make the examples utils
